+++ petfinder/src/main/resources/templates/register.html
@@ -108,51 +108,52 @@
             <div class="card-body">
               <!-- INSERCIÓN DE TH PARA VALIDACIÓN DE REGISTRO -->
               <div th:if=${param.exito}>
-                  <div class = "alert alert-info">Register Success</div>
-              </div>
+                  <div class = "alert alert-info">Register Success</div></div>
 
                     <!-- Definición de register html y Register object backend -->
 
               <form th:action="@{/register}" method="post" th:object="${Register}">
                 <!-- No se puso div class= form group -->
                 <!-- Input name -->
-                <div class="mb-4">
+
+
+                <div class="mb-4 form-group">
                   <div class="input-group has-validation">
                     <span id="txtrg"  class="input-group-text">&#128100</span>
-                    <input id="txtrg" type="text" class="form-control" th:field="*{name}" placeholder="Name" required>
+                    <input id="txtrg" type="text" class="form-control" field="*{name}" name="name" placeholder="Name" required>
                   <div class="invalid-feedback">
                       Your Name is required.
                     </div>
                   </div>
                 </div>
                 <!-- Input lastname -->
-                <div class="mb-4">
+                <div class="mb-4 form-group">
              
                   <div class="input-group has-validation">
                     <span id="txtrg"class="input-group-text">&#128100</span>
-                    <input id="txtrg" type="text" class="form-control" th:field="*{lastname}"  placeholder="Last Name" required>
+                    <input id="txtrg" type="text" class="form-control" field="*{lastname}" name="lastname" placeholder="Last Name" required>
                   <div class="invalid-feedback">
                       Your Last Name is required.
                     </div>
                   </div>
                 
                 </div>
-                <div class="mb-4">
+                <div class="mb-4 form-group">
                   <!-- Input Email -->
                   <!-- Cambio tipo text a email -->
                   <div class="input-group has-validation">
                     <span id="txtrg" class="input-group-text">&#9993</span>
-                    <input id="txtrg" type="email" class="form-control" th:field="*{email}" placeholder="Email" required>
+                    <input id="txtrg" type="email" class="form-control" field="*{email}" name="email" placeholder="Email" required>
                   <div class="invalid-feedback">
                       Your Email is required.
                     </div>
                   </div>   
                   <br>
                   <!-- Input Password -->
-                  <div class="mb-4">
+                  <div class="mb-4 form-group">
                     <div class="input-group has-validation">
                       <span id="txtrg" class="input-group-text">&#128477 </span>
-                      <input id="txtrg" type="password" class="form-control" th:field="*{password}" placeholder="Password" required>
+                      <input id="txtrg" type="password" class="form-control" field="*{password}" name="password"placeholder="Password" required>
                     <div class="invalid-feedback">
                         Your Password is required.
                       </div>
@@ -161,8 +162,8 @@
 
 
                   </div>
-                <div class="d-grid">
-                  <button id="buttonreg" type="submit" class="btn text-light btn-primary"onclick="location.href='#'">Create Account</button>
+                <div class="d-grid form-group">
+                  <button id="buttonreg" type="submit" class="btn btn-primary text-light"onclick="location.href='#'">Create Account</button>
                 </div>
                 <br>
                 <div class="d-grid">
